spring-homework-2

+++ src/main/java/nastyazhabko/dev/properties/AccountProperties.java
@@ -3,21 +3,23 @@ package nastyazhabko.dev.properties;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class AccountProperties {
-    private final int defaultAmount;
-    private final int defaultCommission;
+    private final BigDecimal defaultAmount;
+    private final BigDecimal defaultCommission;
 
-    public AccountProperties(@Value("${account.default-amount}") int defaultAmount, @Value("${account.transfer-commission}") int defaultCommission) {
+    public AccountProperties(@Value("${account.default-amount}") BigDecimal defaultAmount, @Value("${account.transfer-commission}") BigDecimal defaultCommission) {
         this.defaultAmount = defaultAmount;
         this.defaultCommission = defaultCommission;
     }
 
-    public long getDefaultAmount() {
+    public BigDecimal getDefaultAmount() {
         return defaultAmount;
     }
 
-    public int getDefaultCommission() {
+    public BigDecimal getDefaultCommission() {
         return defaultCommission;
     }
 }
